Update of the node server
Model is working. Multiple attributes available, a constructor, and 4
methods (create, read, update and delete)
Add slid.router.js with 2 GET methods and 1 POST.
Add slid.controller.js, it interact between the router and the model in
order to get the list of the slids available, create slids or read
slids.